feat: redesign organisation layout
